hmmap_mod/core: DAX leverage struct page
To fix some consistency bugs I decided to require the backend to now provide
a struct page for a page we wish to map in a DAX manner. This allows us to use
the xarray struct to maintain consistency between DAX and non-DAX accessed
pages. This has been experimentally tested, but I need to look through this in
more detail to make sure I am not missing a case.

Signed-off-by: Adam Manzanares <adam.manzanares@wdc.com>