hmmap_mod/core: Multiple device support
This commit adds support for multiple devices in the core. This required
adding private data in the device for the backend as well as the cache.
The backend and cache apis have been modified to use the device to pass the
private data back to the backend or cache.

In addition the sysfs code had to be reworked to support multiple devices. This
needs more testing before merging into the master.

Signed-off-by: Adam Manzanares <adam.manzanares@wdc.com>